rsconcept/backend/requirements.txt
tzdata
django
djangorestframework
django-cors-headers
django-filter
psycopg2-binary
gunicorn
coreapi